+++ packages/tsurlfilter/src/common/constants.ts
@@ -15,3 +15,12 @@ export const WILDCARD = '*';
 export const LF = '\n';
 export const CR = '\r';
 export const FF = '\f';
+
+export const TAB = '\t';
+
+export const RE_NUMBER = /^\d+$/;
+
+/**
+ * Prefix for ruleset name.
+ */
+export const RULESET_NAME_PREFIX = 'ruleset_';