+++ packages/tswebextension/test/lib/common/storage/extension-storage-decorator.test.ts
@@ -7,6 +7,7 @@ import {
 } from 'vitest';
 import browser from 'webextension-polyfill';
 
+import { BrowserStorage } from '../../../../src/lib/common/storage/core';
 import { createExtensionStorageDecorator, ExtensionStorage } from '../../../../src/lib/common/storage';
 
 describe('createExtensionStorageDecorator', () => {
@@ -15,7 +16,7 @@ describe('createExtensionStorageDecorator', () => {
     let storage: ExtensionStorage<typeof data>;
 
     beforeAll(async () => {
-        storage = new ExtensionStorage(key, browser.storage.local);
+        storage = new ExtensionStorage(key, new BrowserStorage<typeof data>(browser.storage.local));
         await storage.init(data);
     });
 
@@ -39,7 +40,7 @@ describe('createExtensionStorageDecorator', () => {
         // Required for test runtime errors
         // @ts-ignore
         expect(() => fieldDecorator({}, { kind: 'method' })).toThrow(
-            'Class member is not auto accessor',
+            'Class member is not an accessor',
         );
     });