+++ packages/tswebextension/test/lib/common/storage/extension-storage.test.ts
@@ -1,12 +1,13 @@
 import { describe, expect, it } from 'vitest';
 import browser from 'webextension-polyfill';
 
+import { BrowserStorage } from '../../../../src/lib/common/storage/core';
 import { ExtensionStorage } from '../../../../src/lib/common/storage';
 
 describe('ExtensionStorage', () => {
     const key = 'test-key';
     const data = { foo: 'bar', baz: 42 };
-    const api = browser.storage.local;
+    const api = new BrowserStorage<typeof data>(browser.storage.local);
 
     it('should initialize the storage', async () => {
         const storage = new ExtensionStorage<typeof data>(key, api);